修复删除下载中的任务导致崩溃的bug
修复番组列表点击查看时，空指针及题图不显示的bug

+++ src/tv/avfun/util/download/DownloadManager.java
@@ -89,6 +89,7 @@ public class DownloadManager {
      * Delete the download job and related files.
      */
     public void deleteDownload(DownloadJob job){
+        if(job == null) throw new NullPointerException("job cannot be null");
         mProvider.removeDownload(job);
         deleteDownloadFile(job);
     }
